moved all entities to entities directory. updated all classes from package level of visibilty to public level of visibility. updated problematic NuGet warning.
