+++ TexasHoldem/BL/BL.csproj
@@ -1,10 +1,10 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>99782cf7-8736-4c1e-949f-c0b962071b0c</ProjectGuid>
+    <ProjectGuid>{99782CF7-8736-4C1E-949F-C0B962071B0C}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>BL</RootNamespace>
@@ -30,24 +30,29 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="System"/>
-    
-    <Reference Include="System.Core"/>
-    <Reference Include="System.Xml.Linq"/>
-    <Reference Include="System.Data.DataSetExtensions"/>
-    
-    
-    <Reference Include="Microsoft.CSharp"/>
-    
-    <Reference Include="System.Data"/>
-    
-    <Reference Include="System.Net.Http"/>
-    
-    <Reference Include="System.Xml"/>
+    <Reference Include="System" />
+    <Reference Include="System.Core" />
+    <Reference Include="System.Xml.Linq" />
+    <Reference Include="System.Data.DataSetExtensions" />
+    <Reference Include="Microsoft.CSharp" />
+    <Reference Include="System.Data" />
+    <Reference Include="System.Net.Http" />
+    <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Class1.cs" />
+    <Compile Include="entities\Game\Card.cs" />
+    <Compile Include="entities\Game\Deck.cs" />
+    <Compile Include="entities\Game\GameLog.cs" />
+    <Compile Include="entities\Game\GamePreferences.cs" />
+    <Compile Include="entities\Game\LeagueTexasHoldemGame.cs" />
+    <Compile Include="entities\Game\Player.cs" />
+    <Compile Include="entities\Game\RegularTexasHoldemGame.cs" />
+    <Compile Include="entities\Game\Spectator.cs" />
+    <Compile Include="entities\Game\TexasHoldemGame.cs" />
+    <Compile Include="entities\Messages\Notification.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="entities\System\GameCenter.cs" />
+    <Compile Include="entities\User\SystemUser.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
- </Project>
+</Project>
