--- TexasHoldem/BL/Game/Player.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BL.Game
-{
-	class Player : Spectator
-	{
-		public int Tokens { get; set; }
-		public Player(int tokens) : base()
-		{
-			Tokens = tokens;
-		}
-	}
-}
