--- TexasHoldem/BL/Game/Spectator.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BL.Game
-{
-	class Spectator
-	{
-		public Spectator() { }
-	}
-}
