+++ TexasHoldem/BL/entities/Game/Deck.cs
@@ -1,12 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace BL.Game
 {
-	class Deck
+	public class Deck
 	{
 		private List<Card> cards;
 		private Random rand;
