+++ TexasHoldem/BL/entities/Game/LeagueTexasHoldemGame.cs
@@ -1,12 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BL.Game
+﻿namespace BL.Game
 {
-	class LeagueTexasHoldemGame : TexasHoldemGame
+	public class LeagueTexasHoldemGame : TexasHoldemGame
 	{
 		public int MinRank { get; }
 		public int MaxRank { get; }
