+++ TexasHoldem/BL/entities/Game/Player.cs
@@ -1,0 +1,11 @@
+﻿namespace BL.Game
+{
+	public class Player : Spectator
+	{
+		public int Tokens { get; set; }
+		public Player(int tokens) : base()
+		{
+			Tokens = tokens;
+		}
+	}
+}
