+++ TexasHoldem/BL/entities/Game/RegularTexasHoldemGame.cs
@@ -1,0 +1,9 @@
+﻿namespace BL.Game
+{
+	public class RegularTexasHoldemGame : TexasHoldemGame
+	{
+		public RegularTexasHoldemGame(int buyInPolicy, Player gameCreator, GamePreferences gamePreferences) :
+				base(buyInPolicy, gameCreator, gamePreferences)
+		{ }
+	}
+}
