+++ TexasHoldem/BL/entities/Game/Spectator.cs
@@ -1,0 +1,7 @@
+﻿namespace BL.Game
+{
+	public class Spectator
+	{
+		public Spectator() { }
+	}
+}
