+++ TexasHoldem/BL/entities/System/GameCenter.cs
@@ -1,13 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TexasHoldem.User;
 
 namespace BL.System
 {
-	class GameCenter : Messages.Notification
+	public class GameCenter : Messages.Notification
 	{
 		private List<Game.TexasHoldemGame> texasHoldemGames;
 		public Game.TexasHoldemGame createGame(int buyInPolicy, Game.Player player, Game.GamePreferences preferences)
